*** empty log message ***

+++ magick/widget.h
@@ -59,7 +59,8 @@ extern MagickExport void
   XInfoWidget(Display *,XWindows *,const char *),
   XListBrowserWidget(Display *,XWindows *,XWindowInfo *,char const **,
     const char *,const char *,char *),
-  XMonitorWidget(Display *,XWindows *,const char *,const size_t,const size_t),
+  XMonitorWidget(Display *,XWindows *,const char *,const unsigned long,
+    const unsigned long),
   XNoticeWidget(Display *,XWindows *,const char *,const char *),
   XTextViewWidget(Display *,const XResourceInfo *,XWindows *,const unsigned int,
     const char *,char const **);
